+++ src/FilterBundle/Event/Request/ProductProjectionSearchRequestEvent.php
@@ -22,13 +22,32 @@ class ProductProjectionSearchRequestEvent extends Event
     private $request;
 
     /**
+     * The optional search keywords
+     *
+     * @var null|string
+     */
+    private $searchKeyword;
+
+    /**
      * ProductProjectionSearchRequestEvent constructor.
      *
      * @param ProductProjectionSearchRequest $request
+     * @param string|null $searchKeyword
      */
-    public function __construct(ProductProjectionSearchRequest $request)
+    public function __construct(ProductProjectionSearchRequest $request, string $searchKeyword = null)
     {
-        $this->setRequest($request);
+        $this->request = $request;
+        $this->searchKeyword = $searchKeyword;
+    }
+
+    /**
+     * Get searchKeyword
+     *
+     * @return null|string
+     */
+    public function getSearchKeyword()
+    {
+        return $this->searchKeyword;
     }
 
     /**